--- cuHacking/Data/Map/Level.swift
@@ -1,15 +1,0 @@
-//
-//  Floor.swift
-//  cuHacking
-//
-//  Created by Santos on 2019-07-20.
-//  Copyright © 2019 cuHacking. All rights reserved.
-//
-
-import Foundation
-enum Level: String, CaseIterable {
-    case one = "1"
-    case two = "2"
-    case three = "3"
-//    case four = "4"
-}
