+++ cuHacking/Extensions /String+Ext.swift
@@ -10,6 +10,13 @@ import Foundation
 import UIKit
 
 extension String {
+    
+    ///  This method formats a string to use the MGL_Match method in mapbox which can be used for create an NSExpression
+    /// - Parameters:
+    ///   - attribute: The property of the feautre in Mapbox eg. roomType
+    ///   - keys:  The keys to look for in the propery eg. washroom
+    ///   - stringFormat:  the format of the string
+    ///   - includeDefault:  indicates whether or not a default value is included 
     static func formatMGLMatchExpression(attribute: String, keys: [String], stringFormat: String, includeDefault: Bool) -> String {
         var format = "MGL_MATCH(\(attribute),"
         for key in keys {
