--- cuHacking/Map/CardState.swift
@@ -1,13 +1,0 @@
-//
-//  CardState.swift
-//  cuHacking
-//
-//  Created by Santos on 2019-07-30.
-//  Copyright © 2019 cuHacking. All rights reserved.
-//
-
-import Foundation
-enum CardState {
-    case expanded
-    case collapsed
-}
