+++ cuHacking/Profile/Domain/UserProfile.swift
@@ -5,88 +5,20 @@
 //  Created by Santos on 2019-11-08.
 //  Copyright © 2019 cuHacking. All rights reserved.
 //
-/*
- 
- {
-     "operation": "get",
-     "status": "success",
-     "data": {
-         "role": "admin",
-         "application": {
-             "basicInfo": {
-                 "gender": "Prefer not to answer",
-                 "firstName": null,
-                 "ethnicity": "Prefer not to answer",
-                 "emergencyPhone": null,
-                 "otherEthnicity": null,
-                 "lastName": null,
-                 "otherGender": null
-             },
-             "profile": {
-                 "resume": false,
-                 "linkedin": null,
-                 "website": null,
-                 "soughtPosition": "Internship (Co-op)",
-                 "github": null
-             },
-             "personalInfo": {
-                 "major": null,
-                 "minor": null,
-                 "expectedGraduation": 2020,
-                 "degree": "Bachelor",
-                 "cityOfOrigin": null,
-                 "tShirtSize": null,
-                 "wantsShuttle": false,
-                 "otherSchool": null,
-                 "school": "Carleton University",
-                 "dietaryRestrictions": {
-                     "other": null,
-                     "glutenFree": false,
-                     "lactoseFree": false,
-                     "nutFree": false,
-                     "vegetarian": false,
-                     "halal": false
-                 }
-             },
-             "terms": {
-                 "privacyPolicy": false,
-                 "under18": false,
-                 "codeOfConduct": false
-             },
-             "skills": {
-                 "challengeStatement": null,
-                 "accomplishmentStatement": null,
-                 "selfTitle": null,
-                 "numHackathons": 0
-             }
-         },
-         "review": {
-             "wave": 3
-         },
-         "email": "santosgagbegnon@gmail.com",
-         "rsvp": {},
-         "appStatus": "unstarted"
-     }
- }
- 
- */
+
 import Foundation
 import UIKit
 struct MagnetonAPIObject {
 
     struct UserProfile: Codable {
-//        let operation: String
-//        let status: String
         let data: MagnetonAPIObject.Data
     }
+    
     struct Data: Codable {
         let role: String
         let color: String
         let application: Application
-//        let review: Review
         let email: String
-//        let rsvp: [String: String]
-//        let appStatus: String
 
         var badgeColor: UIColor {
             switch color {
@@ -103,13 +35,12 @@ struct MagnetonAPIObject {
             }
         }
     }
+    
     struct Application: Codable {
         let basicInfo: BasicInfo
-//        let profile: Profile
         let personalInfo: PersonalInfo
-//        let terms: Terms
-//        let skills: Skills
     }
+    
     struct BasicInfo: Codable {
         let gender: String?
         let firstName: String?
@@ -128,13 +59,6 @@ struct MagnetonAPIObject {
     }
 
     struct PersonalInfo: Codable {
-//        let major: String?
-//        let minor: String?
-//        let expectedGraduation: Int
-//        let degree: String?
-//        let cityOfOrigin: String?
-//        let tShirtSize: String?
-//        let wantsShuttle: Bool
         let school: String?
         let dietaryRestrictions: DietaryRestrictions
     }
