--- cuHacking/Resources/Colors.xcassets/Blue1.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "colors" : [
-    {
-      "idiom" : "universal",
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "red" : "0x97",
-          "alpha" : "1.000",
-          "blue" : "0xD8",
-          "green" : "0x56"
-        }
-      }
-    }
-  ]
-}
