--- cuHacking/Resources/Colors.xcassets/Blue2.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "colors" : [
-    {
-      "idiom" : "universal",
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "red" : "0xF6",
-          "alpha" : "1.000",
-          "blue" : "0x00",
-          "green" : "0xD5"
-        }
-      }
-    }
-  ]
-}
