--- cuHacking/Resources/Colors.xcassets/Blue3.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "colors" : [
-    {
-      "idiom" : "universal",
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "red" : "0xDD",
-          "alpha" : "1.000",
-          "blue" : "0x1E",
-          "green" : "0x00"
-        }
-      }
-    }
-  ]
-}
