--- cuHacking/Resources/Colors.xcassets/Blue5.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "colors" : [
-    {
-      "idiom" : "universal",
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "red" : "0x00",
-          "alpha" : "1.000",
-          "blue" : "0xA9",
-          "green" : "0x6C"
-        }
-      }
-    }
-  ]
-}
