--- cuHacking/Resources/Colors.xcassets/Blue6.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "colors" : [
-    {
-      "idiom" : "universal",
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "red" : "0xAA",
-          "alpha" : "1.000",
-          "blue" : "0xAA",
-          "green" : "0xAA"
-        }
-      }
-    }
-  ]
-}
