--- cuHacking/UI/CUButton.swift
@@ -1,9 +1,0 @@
-//
-//  CUButton.swift
-//  cuHacking
-//
-//  Created by Santos on 2019-09-02.
-//  Copyright © 2019 cuHacking. All rights reserved.
-//
-
-import Foundation
