--- cuHacking/UI/CUView.swift
@@ -1,10 +1,0 @@
-//
-//  CUView.swift
-//  cuHacking
-//
-//  Created by Santos on 2019-06-28.
-//  Copyright © 2019 cuHacking. All rights reserved.
-//
-
-import UIKit
-class CUView: UIView {}
