--- cuHacking/UI/CUViewController.swift
@@ -1,21 +1,0 @@
-//
-//  CUViewController.swift
-//  cuHacking
-//
-//  Created by Santos on 2019-06-28.
-//  Copyright © 2019 cuHacking. All rights reserved.
-//
-
-import UIKit
-class CUViewController: UIViewController {
-    override func loadView() {
-        let view = CUView()
-        self.view = view
-        super.loadView()
-    }
-
-    override func viewDidLoad() {
-        navigationItem.backBarButtonItem = UIBarButtonItem(title: "", style: .plain, target: nil, action: nil)
-        super.viewDidLoad()
-    }
-}
